+++ central/graphql/resolvers/images.go
@@ -118,36 +118,6 @@ func init() {
 		schema.AddQuery("images(query: String, pagination: Pagination): [Image!]!"),
 		schema.AddQuery("imageCount(query: String): Int!"),
 		schema.AddEnumType("ImageWatchStatus", imageWatchStatuses),
-		schema.AddExtraResolvers("ImageV2", []string{
-			"deploymentCount(query: String): Int!",
-			"deployments(query: String, pagination: Pagination): [Deployment!]!",
-			"imageComponentCount(query: String): Int!",
-			"imageComponents(query: String, pagination: Pagination): [ImageComponent!]!",
-			"imageCVECountBySeverity(query: String): ResourceCountByCVESeverity!",
-			"imageVulnerabilityCount(query: String): Int!",
-			"imageVulnerabilityCounter(query: String): VulnerabilityCounter!",
-			"imageVulnerabilities(query: String, scopeQuery: String, pagination: Pagination): [ImageVulnerability]!",
-			"plottedImageVulnerabilities(query: String): PlottedImageVulnerabilities!",
-			"scan: ImageScan",
-			"topImageVulnerability(query: String): ImageVulnerability",
-			"unusedVarSink(query: String): Int",
-			"watchStatus: ImageWatchStatus!",
-
-			// Image scan-related fields
-			"dataSource: DataSource",
-			"scanNotes: [ImageScan_Note!]!",
-			"operatingSystem: String!",
-			"scanTime: Time",
-			"scannerVersion: String!",
-
-			// Image signature-related fields.
-			"signatureCount: Int!",
-		}),
-		schema.AddQuery("imageV2(id: ID!): ImageV2"),
-		schema.AddQuery("fullImageV2(id: ID!): ImageV2"),
-		schema.AddQuery("imageV2s(query: String, pagination: Pagination): [ImageV2!]!"),
-		schema.AddQuery("imageV2Count(query: String): Int!"),
-		schema.AddEnumType("ImageV2WatchStatus", imageWatchStatuses),
 	)
 }
 
@@ -177,29 +147,6 @@ func (resolver *Resolver) Images(ctx context.Context, args PaginatedQuery) ([]Im
 	return res, err
 }
 
-// ImagesV2 returns GraphQL resolvers for all images using the ImageV2 model
-func (resolver *Resolver) ImageV2s(ctx context.Context, args PaginatedQuery) ([]ImageResolver, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Root, "ImageV2s")
-	if err := readImages(ctx); err != nil {
-		return nil, err
-	}
-	q, err := args.AsV1QueryOrEmpty()
-	if err != nil {
-		return nil, err
-	}
-	imageLoader, err := loaders.GetImageV2Loader(ctx)
-	if err != nil {
-		return nil, err
-	}
-	images, err := imageLoader.FromQuery(ctx, q)
-	resolvers, err := resolver.wrapImageV2sWithContext(ctx, images, err)
-	res := make([]ImageResolver, 0, len(resolvers))
-	for _, resolver := range resolvers {
-		res = append(res, resolver)
-	}
-	return res, err
-}
-
 // ImageCount returns count of all images across deployments
 func (resolver *Resolver) ImageCount(ctx context.Context, args RawQuery) (int32, error) {
 	if features.FlattenImageData.Enabled() {
@@ -215,23 +162,6 @@ func (resolver *Resolver) ImageCount(ctx context.Context, args RawQuery) (int32,
 		return 0, err
 	}
 	imageLoader, err := loaders.GetImageLoader(ctx)
-	if err != nil {
-		return 0, err
-	}
-	return imageLoader.CountFromQuery(ctx, q)
-}
-
-// ImageV2Count returns count of all images across deployments using the ImageV2 model
-func (resolver *Resolver) ImageV2Count(ctx context.Context, args RawQuery) (int32, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Root, "ImageV2Count")
-	if err := readImages(ctx); err != nil {
-		return 0, err
-	}
-	q, err := args.AsV1QueryOrEmpty()
-	if err != nil {
-		return 0, err
-	}
-	imageLoader, err := loaders.GetImageV2Loader(ctx)
 	if err != nil {
 		return 0, err
 	}
@@ -256,21 +186,6 @@ func (resolver *Resolver) Image(ctx context.Context, args struct{ ID graphql.ID 
 	return resolver.wrapImageWithContext(ctx, image, image != nil, err)
 }
 
-// ImageV2 returns a graphql resolver for the identified image, if it exists
-func (resolver *Resolver) ImageV2(ctx context.Context, args struct{ ID graphql.ID }) (ImageResolver, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Root, "ImageV2")
-	if err := readImages(ctx); err != nil {
-		return nil, err
-	}
-
-	imageLoader, err := loaders.GetImageV2Loader(ctx)
-	if err != nil {
-		return nil, err
-	}
-	image, err := imageLoader.FromID(ctx, string(args.ID))
-	return resolver.wrapImageV2WithContext(ctx, image, image != nil, err)
-}
-
 // FullImage returns a graphql resolver for the identified image, if it exists
 func (resolver *Resolver) FullImage(ctx context.Context, args struct{ ID graphql.ID }) (ImageResolver, error) {
 	if features.FlattenImageData.Enabled() {
@@ -289,29 +204,8 @@ func (resolver *Resolver) FullImage(ctx context.Context, args struct{ ID graphql
 	return resolver.wrapImageWithContext(ctx, image, image != nil, err)
 }
 
-// FullImage returns a graphql resolver for the identified image, if it exists
-func (resolver *Resolver) FullImageV2(ctx context.Context, args struct{ ID graphql.ID }) (ImageResolver, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Root, "FullImageV2")
-	if err := readImages(ctx); err != nil {
-		return nil, err
-	}
-
-	imageLoader, err := loaders.GetImageV2Loader(ctx)
-	if err != nil {
-		return nil, err
-	}
-	image, err := imageLoader.FullImageWithID(ctx, string(args.ID))
-	return resolver.wrapImageV2WithContext(ctx, image, image != nil, err)
-}
-
 // Deployments returns the deployments which use this image for the identified image, if it exists
 func (resolver *imageResolver) Deployments(ctx context.Context, args PaginatedQuery) ([]*deploymentResolver, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "Deployments")
-	return resolver.root.Deployments(resolver.withImageScopeContext(ctx), args)
-}
-
-// Deployments returns the deployments which use this image for the identified image, if it exists
-func (resolver *imageV2Resolver) Deployments(ctx context.Context, args PaginatedQuery) ([]*deploymentResolver, error) {
 	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "Deployments")
 	return resolver.root.Deployments(resolver.withImageScopeContext(ctx), args)
 }
@@ -322,20 +216,8 @@ func (resolver *imageResolver) DeploymentCount(ctx context.Context, args RawQuer
 	return resolver.root.DeploymentCount(resolver.withImageScopeContext(ctx), args)
 }
 
-// DeploymentCount returns the number of deployments which use this image for the identified image, if it exists
-func (resolver *imageV2Resolver) DeploymentCount(ctx context.Context, args RawQuery) (int32, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "DeploymentCount")
-	return resolver.root.DeploymentCount(resolver.withImageScopeContext(ctx), args)
-}
-
 // TopImageVulnerability returns the image vulnerability with the top CVSS score.
 func (resolver *imageResolver) TopImageVulnerability(ctx context.Context, args RawQuery) (ImageVulnerabilityResolver, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "TopImageVulnerability")
-	return resolver.root.TopImageVulnerability(resolver.withImageScopeContext(ctx), args)
-}
-
-// TopImageVulnerability returns the image vulnerability with the top CVSS score.
-func (resolver *imageV2Resolver) TopImageVulnerability(ctx context.Context, args RawQuery) (ImageVulnerabilityResolver, error) {
 	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "TopImageVulnerability")
 	return resolver.root.TopImageVulnerability(resolver.withImageScopeContext(ctx), args)
 }
@@ -346,36 +228,12 @@ func (resolver *imageResolver) ImageVulnerabilities(ctx context.Context, args Pa
 	return resolver.root.ImageVulnerabilities(resolver.withImageScopeContext(ctx), args)
 }
 
-// ImageVulnerabilities returns, as ImageVulnerabilityResolver, the vulnerabilities for the image
-func (resolver *imageV2Resolver) ImageVulnerabilities(ctx context.Context, args PaginatedQuery) ([]ImageVulnerabilityResolver, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "ImageVulnerabilities")
-	return resolver.root.ImageVulnerabilities(resolver.withImageScopeContext(ctx), args)
-}
-
 func (resolver *imageResolver) ImageVulnerabilityCount(ctx context.Context, args RawQuery) (int32, error) {
 	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "ImageVulnerabilityCount")
 	return resolver.root.ImageVulnerabilityCount(resolver.withImageScopeContext(ctx), args)
 }
 
-func (resolver *imageV2Resolver) ImageVulnerabilityCount(ctx context.Context, args RawQuery) (int32, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "ImageVulnerabilityCount")
-	return resolver.root.ImageVulnerabilityCount(resolver.withImageScopeContext(ctx), args)
-}
-
 func (resolver *imageResolver) ImageVulnerabilityCounter(ctx context.Context, args RawQuery) (*VulnerabilityCounterResolver, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "ImageVulnerabilityCounter")
-	// In this function, the resolver is obtained by a call to either Image or Images from the root resolver,
-	// applying scoped access control to avoid exposing images the requester should not be aware of.
-	// The image ID is added to the context (by withImageScopeContext) to restrict the vulnerability search
-	// to the CVEs linked to the image.
-	// If no context elevation is done, then scoped access control is applied again on top of the image ID filtering
-	// leading to additional table joins in DB and poor performance.
-	// The context is elevated to bypass the scoped access control and improve the performance,
-	// considering the fact that the image ID was obtained by applying scoped access control rules.
-	return resolver.root.ImageVulnerabilityCounter(resolver.withElevatedImageScopeContext(ctx), args)
-}
-
-func (resolver *imageV2Resolver) ImageVulnerabilityCounter(ctx context.Context, args RawQuery) (*VulnerabilityCounterResolver, error) {
 	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "ImageVulnerabilityCounter")
 	// In this function, the resolver is obtained by a call to either Image or Images from the root resolver,
 	// applying scoped access control to avoid exposing images the requester should not be aware of.
@@ -402,26 +260,7 @@ func (resolver *imageResolver) ImageCVECountBySeverity(ctx context.Context, q Ra
 	return resolver.root.wrapResourceCountByCVESeverityWithContext(ctx, val, err)
 }
 
-func (resolver *imageV2Resolver) ImageCVECountBySeverity(ctx context.Context, q RawQuery) (*resourceCountBySeverityResolver, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "ImageCVECountBySeverity")
-
-	if err := readImages(ctx); err != nil {
-		return nil, err
-	}
-	query, err := q.AsV1QueryOrEmpty()
-	if err != nil {
-		return nil, err
-	}
-	val, err := resolver.root.ImageCVEView.CountBySeverity(resolver.withImageScopeContext(ctx), query)
-	return resolver.root.wrapResourceCountByCVESeverityWithContext(ctx, val, err)
-}
-
 func (resolver *imageResolver) ImageComponents(ctx context.Context, args PaginatedQuery) ([]ImageComponentResolver, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "ImageComponents")
-	return resolver.root.ImageComponents(resolver.withImageScopeContext(ctx), args)
-}
-
-func (resolver *imageV2Resolver) ImageComponents(ctx context.Context, args PaginatedQuery) ([]ImageComponentResolver, error) {
 	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "ImageComponents")
 	return resolver.root.ImageComponents(resolver.withImageScopeContext(ctx), args)
 }
@@ -429,17 +268,6 @@ func (resolver *imageV2Resolver) ImageComponents(ctx context.Context, args Pagin
 func (resolver *imageResolver) ImageComponentCount(ctx context.Context, args RawQuery) (int32, error) {
 	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "ImageComponentCount")
 	return resolver.root.ImageComponentCount(resolver.withImageScopeContext(ctx), args)
-}
-
-func (resolver *imageV2Resolver) ImageComponentCount(ctx context.Context, args RawQuery) (int32, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "ImageComponentCount")
-	return resolver.root.ImageComponentCount(resolver.withImageScopeContext(ctx), args)
-}
-
-func (resolver *imageV2Resolver) Names(ctx context.Context) ([]*imageNameResolver, error) {
-	resolver.ensureData(ctx)
-	value := []*storage.ImageName{resolver.data.GetName()}
-	return resolver.root.wrapImageNames(value, nil)
 }
 
 func (resolver *imageResolver) withImageScopeContext(ctx context.Context) context.Context {
@@ -458,33 +286,7 @@ func (resolver *imageResolver) withImageScopeContext(ctx context.Context) contex
 	})
 }
 
-func (resolver *imageV2Resolver) withImageScopeContext(ctx context.Context) context.Context {
-	if ctx == nil {
-		err := utils.ShouldErr(errors.New("argument 'ctx' is nil"))
-		if err != nil {
-			log.Error(err)
-		}
-	}
-	if resolver.ctx == nil {
-		resolver.ctx = ctx
-	}
-	return scoped.Context(resolver.ctx, scoped.Scope{
-		Level: v1.SearchCategory_IMAGES_V2,
-		IDs:   []string{resolver.data.GetId()},
-	})
-}
-
 func (resolver *imageResolver) withElevatedImageScopeContext(ctx context.Context) context.Context {
-	return sac.WithGlobalAccessScopeChecker(
-		resolver.withImageScopeContext(ctx),
-		sac.AllowFixedScopes(
-			sac.AccessModeScopeKeys(storage.Access_READ_ACCESS),
-			sac.ResourceScopeKeys(resources.Image),
-		),
-	)
-}
-
-func (resolver *imageV2Resolver) withElevatedImageScopeContext(ctx context.Context) context.Context {
 	return sac.WithGlobalAccessScopeChecker(
 		resolver.withImageScopeContext(ctx),
 		sac.AllowFixedScopes(
@@ -506,55 +308,13 @@ func (resolver *Resolver) getImage(ctx context.Context, id string) *storage.Imag
 	return image
 }
 
-func (resolver *Resolver) getImageV2(ctx context.Context, id string) *storage.ImageV2 {
-	imageLoader, err := loaders.GetImageV2Loader(ctx)
-	if err != nil {
-		return nil
-	}
-	image, err := imageLoader.FromID(ctx, id)
-	if err != nil {
-		return nil
-	}
-	return image
-}
-
 // PlottedImageVulnerabilities returns the data required by top risky entity scatter-plot on vuln mgmt dashboard
 func (resolver *imageResolver) PlottedImageVulnerabilities(ctx context.Context, args RawQuery) (*PlottedImageVulnerabilitiesResolver, error) {
 	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "PlottedImageVulnerabilities")
 	return resolver.root.PlottedImageVulnerabilities(resolver.withImageScopeContext(ctx), args)
 }
 
-// PlottedImageVulnerabilities returns the data required by top risky entity scatter-plot on vuln mgmt dashboard
-func (resolver *imageV2Resolver) PlottedImageVulnerabilities(ctx context.Context, args RawQuery) (*PlottedImageVulnerabilitiesResolver, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "PlottedImageVulnerabilities")
-	return resolver.root.PlottedImageVulnerabilities(resolver.withImageScopeContext(ctx), args)
-}
-
 func (resolver *imageResolver) Scan(ctx context.Context) (*imageScanResolver, error) {
-	resolver.ensureData(ctx)
-
-	// If scan is pulled, it is most likely to fetch all components and vulns contained in image.
-	// Therefore, load the image again with full scan.
-	imageLoader, err := loaders.GetImageLoader(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	image, err := imageLoader.FullImageWithID(ctx, resolver.data.GetId())
-	if err != nil {
-		return nil, err
-	}
-	scan := image.GetScan()
-
-	res, err := resolver.root.wrapImageScan(scan, true, nil)
-	if err != nil || res == nil {
-		return nil, err
-	}
-	res.ctx = resolver.withImageScopeContext(ctx)
-	return res, nil
-}
-
-func (resolver *imageV2Resolver) Scan(ctx context.Context) (*imageScanResolver, error) {
 	resolver.ensureData(ctx)
 
 	// If scan is pulled, it is most likely to fetch all components and vulns contained in image.
@@ -596,29 +356,7 @@ func (resolver *imageResolver) WatchStatus(ctx context.Context) (string, error) 
 	return notWatchedImageWatchStatus, nil
 }
 
-func (resolver *imageV2Resolver) WatchStatus(ctx context.Context) (string, error) {
-	defer metrics.SetGraphQLOperationDurationTime(time.Now(), pkgMetrics.Images, "WatchStatus")
-	if err := readAuth(resources.WatchedImage)(ctx); err != nil {
-		if errors.Is(err, errox.NotAuthorized) {
-			return unknownImageWatchStatus, nil
-		}
-		return "", err
-	}
-	watched, err := resolver.root.WatchedImageDataStore.Exists(ctx, resolver.data.GetName().GetFullName())
-	if err != nil {
-		return "", err
-	}
-	if watched {
-		return watchedImageStatus, nil
-	}
-	return notWatchedImageWatchStatus, nil
-}
-
 func (resolver *imageResolver) UnusedVarSink(_ context.Context, _ RawQuery) *int32 {
-	return nil
-}
-
-func (resolver *imageV2Resolver) UnusedVarSink(_ context.Context, _ RawQuery) *int32 {
 	return nil
 }
 
@@ -629,17 +367,7 @@ func (resolver *imageResolver) DataSource(_ context.Context) (*dataSourceResolve
 	return resolver.root.wrapDataSource(value, true, nil)
 }
 
-func (resolver *imageV2Resolver) DataSource(_ context.Context) (*dataSourceResolver, error) {
-	value := resolver.data.GetScan().GetDataSource()
-	return resolver.root.wrapDataSource(value, true, nil)
-}
-
 func (resolver *imageResolver) ScanNotes(_ context.Context) []string {
-	value := resolver.data.GetScan().GetNotes()
-	return stringSlice(value)
-}
-
-func (resolver *imageV2Resolver) ScanNotes(_ context.Context) []string {
 	value := resolver.data.GetScan().GetNotes()
 	return stringSlice(value)
 }
@@ -649,17 +377,7 @@ func (resolver *imageResolver) OperatingSystem(_ context.Context) string {
 	return value
 }
 
-func (resolver *imageV2Resolver) OperatingSystem(_ context.Context) string {
-	value := resolver.data.GetScan().GetOperatingSystem()
-	return value
-}
-
 func (resolver *imageResolver) ScanTime(_ context.Context) (*graphql.Time, error) {
-	value := resolver.data.GetScan().GetScanTime()
-	return protocompat.ConvertTimestampToGraphqlTimeOrError(value)
-}
-
-func (resolver *imageV2Resolver) ScanTime(_ context.Context) (*graphql.Time, error) {
 	value := resolver.data.GetScan().GetScanTime()
 	return protocompat.ConvertTimestampToGraphqlTimeOrError(value)
 }
@@ -669,27 +387,7 @@ func (resolver *imageResolver) ScannerVersion(_ context.Context) string {
 	return value
 }
 
-func (resolver *imageV2Resolver) ScannerVersion(_ context.Context) string {
-	value := resolver.data.GetScan().GetScannerVersion()
-	return value
-}
-
 func (resolver *imageResolver) SignatureCount(ctx context.Context) (int32, error) {
-	resolver.ensureData(ctx)
-
-	imageLoader, err := loaders.GetImageLoader(ctx)
-	if err != nil {
-		return 0, err
-	}
-
-	image, err := imageLoader.FullImageWithID(ctx, resolver.data.GetId())
-	if err != nil {
-		return 0, err
-	}
-	return int32(len(image.GetSignature().GetSignatures())), nil
-}
-
-func (resolver *imageV2Resolver) SignatureCount(ctx context.Context) (int32, error) {
 	resolver.ensureData(ctx)
 
 	imageLoader, err := loaders.GetImageLoader(ctx)
@@ -713,6 +411,7 @@ func (resolver *imageResolver) ComponentCount(_ context.Context) int32 {
 }
 
 func (resolver *imageResolver) CriticalCveCount(_ context.Context) int32 {
+	// This only exists for conformity to the common interface, these counts do not exist in the original image data model, and thus return 0
 	return 0
 }
 
@@ -725,6 +424,7 @@ func (resolver *imageResolver) CveCount(_ context.Context) int32 {
 }
 
 func (resolver *imageResolver) FixableCriticalCveCount(_ context.Context) int32 {
+	// This only exists for conformity to the common interface, these counts do not exist in the original image data model, and thus return 0
 	return 0
 }
 
@@ -737,34 +437,42 @@ func (resolver *imageResolver) FixableCveCount(_ context.Context) int32 {
 }
 
 func (resolver *imageResolver) FixableImportantCveCount(_ context.Context) int32 {
+	// This only exists for conformity to the common interface, these counts do not exist in the original image data model, and thus return 0
 	return 0
 }
 
 func (resolver *imageResolver) FixableLowCveCount(_ context.Context) int32 {
+	// This only exists for conformity to the common interface, these counts do not exist in the original image data model, and thus return 0
 	return 0
 }
 
 func (resolver *imageResolver) FixableModerateCveCount(_ context.Context) int32 {
+	// This only exists for conformity to the common interface, these counts do not exist in the original image data model, and thus return 0
 	return 0
 }
 
 func (resolver *imageResolver) FixableUnknownCveCount(_ context.Context) int32 {
+	// This only exists for conformity to the common interface, these counts do not exist in the original image data model, and thus return 0
 	return 0
 }
 
 func (resolver *imageResolver) ImportantCveCount(_ context.Context) int32 {
+	// This only exists for conformity to the common interface, these counts do not exist in the original image data model, and thus return 0
 	return 0
 }
 
 func (resolver *imageResolver) LowCveCount(_ context.Context) int32 {
+	// This only exists for conformity to the common interface, these counts do not exist in the original image data model, and thus return 0
 	return 0
 }
 
 func (resolver *imageResolver) ModerateCveCount(_ context.Context) int32 {
+	// This only exists for conformity to the common interface, these counts do not exist in the original image data model, and thus return 0
 	return 0
 }
 
 func (resolver *imageResolver) UnknownCveCount(_ context.Context) int32 {
+	// This only exists for conformity to the common interface, these counts do not exist in the original image data model, and thus return 0
 	return 0
 }
